requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
dnspython==2.8.0
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyJWT==2.10.1
pymongo==4.15.5
python-dotenv==1.2.1
Werkzeug==3.1.4
